Move sources under src directory
Update references as required.
