+++ packages/codegen.go/src/responses.ts
@@ -5,7 +5,7 @@
 
 import { comment } from '@azure-tools/codegen';
 import { values } from '@azure-tools/linq';
-import * as go from '../codemodel.go/gocodemodel.js';
+import * as go from '../../codemodel.go/src/gocodemodel.js';
 import { commentLength, contentPreamble, sortAscending } from './helpers.js';
 import { ImportManager } from './imports.js';
 import { getStar } from './models.js';
