+++ packages/naming.go/src/mappings.ts
