+++ packages/naming.go/src/naming.ts
